+++ src/main/java/com/pomina/erpapp/systemconfigsecurity/sysservice/RedisTokenService.java
@@ -6,7 +6,6 @@ import org.apache.commons.codec.digest.DigestUtils;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
-import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Objects;
@@ -17,13 +16,13 @@ public class RedisTokenService {
 
     private final RedisTemplate<String, RefreshTokenInfo> redisTemplate;
 
-    public RefreshTokenInfo getKey(Long userId, String deviceId, String userAgent) {
+    public RefreshTokenInfo getKey(Integer userId, String deviceId, String userAgent) {
         String key = buildKey(userId, deviceId, userAgent);
         return redisTemplate.opsForValue()
                 .get(key);
     }
 
-    public void saveToken(Long userId, String deviceId, String refreshToken, String userAgent, long expiryMs) {
+    public void saveToken(Integer userId, String deviceId, String refreshToken, String userAgent, long expiryMs) {
         String key = buildKey(userId, deviceId, userAgent);
         RefreshTokenInfo info = RefreshTokenInfo.builder()
                 .encryptedToken(encrypt(refreshToken))
@@ -35,7 +34,7 @@ public class RedisTokenService {
                 .set(key, info, Duration.ofMillis(expiryMs));
     }
 
-    public boolean isValidRefreshToken(Long userId, String deviceId, String refreshToken, String userAgent) {
+    public boolean isValidRefreshToken(Integer userId, String deviceId, String refreshToken, String userAgent) {
         RefreshTokenInfo info = redisTemplate.opsForValue()
                 .get(buildKey(userId, deviceId, userAgent));
         return info != null &&
@@ -43,7 +42,7 @@ public class RedisTokenService {
                 Objects.equals(info.getUserAgent(), userAgent);
     }
 
-    public void deleteToken(Long userId, String deviceId, String userAgent) {
+    public void deleteToken(Integer userId, String deviceId, String userAgent) {
         redisTemplate.delete(buildKey(userId, deviceId, userAgent));
     }
 
@@ -51,7 +50,7 @@ public class RedisTokenService {
         return DigestUtils.sha256Hex(raw);
     }
 
-    private String buildKey(Long userId, String deviceId, String userAgent) {
+    private String buildKey(Integer userId, String deviceId, String userAgent) {
         return "refresh_token:" + userId + ":" + deviceId + ":" + userAgent;
     }
 }
